requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.4.2
matplotlib==3.8.4
mlflow==2.12.2
prometheus-client==0.20.0
requests==2.31.0
pytest==8.2.0